refactor: merge FileUtils and FileNameUtils classes
