+++ .gitignore
@@ -396,3 +396,6 @@ stats.csv
 
 ## Storybook
 **/storybook-static
+
+## AI settings
+.claude/settings.local.json
