--- frontend/libs/studio-pure-functions/src/FileNameUtils/index.ts
@@ -1,3 +1,0 @@
-export { FileNameUtils } from './FileNameUtils';
-export { FileNameErrorResult } from './FileNameUtils';
-export { FILE_NAME_REGEX } from './FileNameUtils';
