+++ frontend/libs/studio-pure-functions/src/FileUtils/index.ts
@@ -1,1 +1,1 @@
-export { FileUtils } from './FileUtils';
+export { FileUtils, FileNameErrorResult, FILE_NAME_REGEX } from './FileUtils';
