+++ frontend/libs/studio-pure-functions/src/index.ts
@@ -1,7 +1,6 @@
 export * from './ArrayUtils';
 export * from './BlobDownloader';
 export * from './DateUtils';
-export * from './FileNameUtils';
 export * from './FileUtils';
 export * from './NumberUtils';
 export * from './ObjectUtils';
