+++ frontend/packages/ux-editor/src/components/config/editModal/EditImage/LocalImage/PreviewImageSummary/PreviewFileInfo.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import classes from './PreviewFileInfo.module.css';
 import { StudioParagraph } from '@studio/components';
-import { FileNameUtils } from '@studio/pure-functions';
+import { FileUtils } from '@studio/pure-functions';
 
 interface PreviewFileInfoProps {
   existingImageUrl: string;
@@ -11,7 +11,7 @@ export const PreviewFileInfo = ({ existingImageUrl }: PreviewFileInfoProps) => {
   return (
     <div className={classes.fileInfoContainer}>
       <StudioParagraph className={classes.fileName} title={existingImageUrl}>
-        {FileNameUtils.extractFileName(existingImageUrl)}
+        {FileUtils.extractFileName(existingImageUrl)}
       </StudioParagraph>
     </div>
   );
